requirements.txt
python-dotenv
pandas
numpy
pymongo
seaborn
matplotlib
scikit-learn
flask
certifi

-e .